+++ app/src/main/java/com/xunda/mo/model/Friend_Details_Bean.java
@@ -65,7 +65,7 @@ public class Friend_Details_Bean {
         @JsonProperty("userNum")
         private Double userNum;
         @JsonProperty("vipType")
-        private Long vipType;
+        private int vipType;
         @JsonProperty("areaName")
         private String areaName;
         @JsonProperty("friendStatus")
